+++ opero/opero/doctype/contract_documents/contract_documents.py
@@ -1,0 +1,5 @@
+from frappe.model.document import Document
+
+
+class ContractDocuments(Document):
+	pass